+++ R/init_bdsreader.R
@@ -21,6 +21,7 @@
 #'   \item{dob}{Date. Date of birth of the child.}
 #'   \item{dobm}{Date. Date of birth of the mother.}
 #'   \item{dobf}{Date. Date of birth of the father.}
+#'   \item{mendate}{Date. Date of menarche (BDS 312), if known.}
 #'   \item{src}{Character. Data source or study identifier.}
 #'   \item{dnr}{Character. Donor or provenance code.}
 #'   \item{sex}{Character. Sex of the child (e.g., "male", "female").}
@@ -65,6 +66,7 @@ init_bdsreader <- function(psn = NULL, xyz = NULL) {
       dob = as.Date(NA),
       dobm = as.Date(NA),
       dobf = as.Date(NA),
+      mendate = as.Date(NA),
       src = NA_character_,
       dnr = NA_character_,
       sex = NA_character_,